Refactor: Rename UIBlocks folder to ui-blocks
