+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 "use client"
 import { ApolloClientProvider } from "~/components/ApolloClientProvider"
 import { GameCard } from "~/components/GameCard"
-import { Layout } from "~/components/UIBlocks/Layout"
-import { LoadingSpinner } from "~/components/UIBlocks/LoadingSpinner"
-import { Pagination } from "~/components/UIBlocks/Pagination"
+import { Layout } from "~/components/ui-blocks/Layout"
+import { LoadingSpinner } from "~/components/ui-blocks/LoadingSpinner"
+import { Pagination } from "~/components/ui-blocks/Pagination"
 import { useGamesQuery } from "~/graphql/generated"
 import { usePage } from "~/hooks/usePage"
 import { useSearchParams } from "next/navigation"
